moved integration test to feature folder where it belongs
